レイアウト&bookmark機能調整

+++ app/controllers/people/bookmarks_controller.rb
@@ -1,21 +1,21 @@
 class People::BookmarksController < ApplicationController
-  
+
   def create
-    team = Team.find(params[:team_id])
-    bookmark = current_person.bookmarks.new(team_id: team.id)
+    @team = Team.find(params[:team_id])
+    bookmark = @team.bookmarks.new(person_id: current_person.id)
     bookmark.save
-    redirect_to people_team_path(team)
+    redirect_to people_team_path(@team)
   end
-  
+
   def destroy
-    team= Team.find(params[:team_id])
-    bookmark = current_person.bookmarks.find_by(team_id: team.id)
-    bookmark.destroy
-    redirect_to people_team_path(team)
+    @team= Team.find(params[:team_id])
+    bookmark = @team.bookmarks.find_by(person_id: current_person.id)
+    if bookmark.present?
+        bookmark.destroy
+        redirect_to people_team_path(@team)
+    else
+        redirect_to people_team_path(@team)
+    end
   end
-  
-  def index
-    @bookmarks = Bookmark.where(team_id: current_person.id)
-  end
-  
+
 end
